Changed float to double in config to prevent crash on death

+++ src/main/java/logictechcorp/reagenchant/core/ReagenchantConfig.java
@@ -40,8 +40,8 @@ public class ReagenchantConfig {
     public static class Common {
         public final ForgeConfigSpec.ConfigValue<Boolean> enableQuarkCompatibility;
         public final ForgeConfigSpec.ConfigValue<Boolean> unbreakableItems;
-        public final ForgeConfigSpec.ConfigValue<Float> percentOfXpDroppedOnDeath;
-        public final ForgeConfigSpec.ConfigValue<Float> percentOfDroppedXpLost;
+        public final ForgeConfigSpec.ConfigValue<Double> percentOfXpDroppedOnDeath;
+        public final ForgeConfigSpec.ConfigValue<Double> percentOfDroppedXpLost;
 
         Common(ForgeConfigSpec.Builder builder) {
             builder.comment("Reagenchant common configuration")
@@ -65,11 +65,11 @@ public class ReagenchantConfig {
             this.percentOfXpDroppedOnDeath = builder
                     .comment("The percent of experience which should drop on death; Default: 0.5")
                     .translation(createTranslation("common.tweaks.player.percent_of_xp_dropped_on_death"))
-                    .defineInRange("percentOfXpDroppedOnDeath", 0.5F, 0.0F, 1.0F, Float.class);
+                    .defineInRange("percentOfXpDroppedOnDeath", 0.5, 0.0, 1.0, Double.class);
             this.percentOfDroppedXpLost = builder
                     .comment("The percent of experience dropped which should be lost; Default: 0.25")
                     .translation(createTranslation("common.tweaks.player.percent_of_dropped_xp_lost"))
-                    .defineInRange("percentOfDroppedXpLost", 0.25F, 0.0F, 1.0F, Float.class);
+                    .defineInRange("percentOfDroppedXpLost", 0.25, 0.0, 1.0, Double.class);
             builder.pop();
             builder.pop();
             builder.pop();
